readnest-mcp/pyproject.toml
[project]
name = "readnest-mcp"
version = "0.1.0"
description = "ReadNest MCP Server - Access your reading data via Model Context Protocol"
authors = [{name = "Yayoi", email = "yayoi@readnest.jp"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
readnest-mcp = "readnest_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/readnest_mcp"]
